+++ src/core/StorageProviders/MongoProvider.js
@@ -20,7 +20,7 @@ exports.ConfigurationDefaults =
 
 //---------------------------------------------------------------------
 exports.NewProvider =
-	function NewProvider( Server, Settings )
+	function NewProvider( Server, Service, Settings )
 	{
 		let storage_provider = {};